Minor file renaming in 32016 Co Pro
