+++ src/Makefile.am
@@ -29,12 +29,12 @@ b_em_SOURCES = \
 	6502tube.c \
 	65816.c \
 	NS32016/32016.c \
+	NS32016/32016_debug.c \
 	NS32016/Decode.c \
 	NS32016/NSDis.c \
 	NS32016/Profile.c \
 	NS32016/Trap.c \
 	NS32016/mem32016.c \
-	NS32016/debug.c \
 	Z80.c \
 	z80dis.c \
 	acia.c \
